requirements.txt
asgiref==3.3.1
Django==3.1.3
django-rest==0.8.6
django-rest-framework==0.1.0
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
gunicorn
psycopg2
django-cors-headers==3.5.0
datetime
PyJWT==1.7.1
pytz==2020.1